feat: add result dialog widget
